+++ lib/generators/modulorails/gitlabci/gitlabci_generator.rb
@@ -1,13 +1,14 @@
 # frozen_string_literal: true
 
-require 'rails/generators'
+require 'modulorails/generators/base'
 
-class Modulorails::GitlabciGenerator < Rails::Generators::Base
+class Modulorails::GitlabciGenerator < Modulorails::Generators::Base
 
-  source_root File.expand_path('templates', __dir__)
   desc 'This generator creates a template for a .gitlab-ci.yml file at root'
 
-  def create_config_file
+  protected
+
+  def create_config
     @data = Modulorails.data
     @image_name = @data.name.parameterize
     @environment_name = @data.environment_name
@@ -25,23 +26,8 @@ class Modulorails::GitlabciGenerator < Rails::Generators::Base
     # Remove the database-ci template if it exists.
     # It used to be referenced by the gitlab-ci template.
     remove_file 'config/database-ci.yml'
-
-    # Create file to avoid this generator on next modulorails launch
-    create_keep_file
   rescue StandardError => e
     warn("[Modulorails] Error: cannot generate CI configuration: #{e.message}")
-  end
-
-  private
-
-  def create_keep_file
-    file = '.modulorails-gitlab-ci'
-
-    # Create file to avoid this generator on next modulorails launch
-    copy_file(file, file)
-
-    say "Add #{file} to git"
-    `git add #{file}`
   end
 
 end