+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file is used to list changes made in each version of the gem.
   - Remove docker-compose.prod.
   - Rename docker-compose.yml to compose.yml.
   - Remove version from compose.yml.
+  - Auto-stop staging container after 7 days.
 - Version Docker generator.
 - Add a generator for project initialization:
   - Add default configuration for production and staging environments.