+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This file is used to list changes made in each version of the gem.
   - Rename docker-compose.yml to compose.yml.
   - Remove version from compose.yml.
   - Auto-stop staging container after 7 days.
+  - Use `rails` chart in Gitlab CI templates.
 - Version Docker generator.
 - Add a generator for project initialization:
   - Add default configuration for production and staging environments.